requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
Faker==37.4.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
